Shipment_Assignment/requirements.txt
django==3.0.2
psycopg2
djangorestframework
requests